feat: 修改主机端数据类型 (通过测试)
DETAILS:
1. 修改 Request 的数据类型, 将温度都修改为 Integer 类型, 如果前端需要展示的话, 除以 100 即可;
2. 主机后端和从机后端的交互温度时的参数类型均改为了 Integer 类型;
3. 调度函数的调度时间改为从配置文件中获取.

+++ master/src/main/java/org/app/controller/SlaveController.java
@@ -84,7 +84,7 @@ public class SlaveController {
      * 接收从机请求, 从机给出的每个参数都不应为 null
      */
     @PostMapping("/OnSlaverPower")
-    public R<String> slaveRequest(Long roomId, double setTemp, double curTemp, String mode) {
+    public R<String> slaveRequest(Long roomId, Integer setTemp, Integer curTemp, String mode) {
         log.info("从机请求参数: {}, {}, {}, {}", roomId, setTemp, curTemp, mode);
         var request = new Request();
         request.setRoomId(roomId);
